Renamed arc_exceptions simply as exceptions
