+++ arc/job/trsh.py
@@ -12,7 +12,7 @@ import os
 import cclib
 import numpy as np
 
-from arc.arc_exceptions import SpeciesError, TrshError
+from arc.exceptions import SpeciesError, TrshError
 from arc.common import get_logger, determine_ess
 from arc.settings import servers, delete_command, list_available_nodes_command, submit_filename, \
     inconsistency_ab, inconsistency_az, maximum_barrier, rotor_scan_resolution
