+++ arc/reaction.py
@@ -13,7 +13,7 @@ from rmgpy.reaction import Reaction
 from arc.common import get_logger
 import arc.rmgdb as rmgdb
 from arc.species.species import ARCSpecies
-from arc.arc_exceptions import ReactionError, InputError
+from arc.exceptions import ReactionError, InputError
 from arc.settings import default_ts_methods
 
 ##################################################################
